agents/requirements.txt
web3
python-dotenv 